Added post editing functionality
Added a button on the top-right of a post page that will let a logged-in
user edit their post. The contents of the post will be autofilled in.
Once the user submits, it will update the post in the database.
The button is put in the same container as the delete button.